Update code to reflect latest changes
